re facted a bunch of person stats reports.
